# swagger/swagger_output.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "oneM2M API",
        "description": "oneM2M API 1.0",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default",
            "variables": {}
        }
    ],
    "paths": {
...
        "/api/onem2m/access-controls": {
            "post": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "CSEBase 리소스 하위 Access Control Policy 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ACP_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "CB 하위 ACP 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/access-controls/{ae}": {
            "post": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "Application Entity 리소스 하위 Access Control Policy 리소스 등록",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더)",
                        "example": "STest_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ACP_Create"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "AE 하위 ACP 리소스 등록",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/access-controls/{acp}": {
            "get": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "CSEBase 리소스 하위 Access Control Policy 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "CB 하위 ACP 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "CSEBase 리소스 하위 Access Control Policy 리소스 업데이트",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ACP_Update"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "CB 하위 ACP 리소스 업데이트",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "CSEBase 리소스 하위 Access Control Policy 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "CB 하위 ACP 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/onem2m/access-controls/{ae}/{acp}": {
            "get": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "Application Entity 리소스 하위 Access Control Policy 리소스 조회",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "AE 하위 ACP 리소스 조회",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "Application Entity 리소스 하위 Access Control Policy 리소스 업데이트",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ACP_Update"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "AE 하위 ACP 리소스 업데이트",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Access Control Policy"
                ],
                "description": "Application Entity 리소스 하위 Access Control Policy 리소스 삭제",
                "parameters": [
                    {
                        "name": "X-M2M-Origin",
                        "in": "header",
                        "description": "리소스의 Origin (헤더), ACP 등록 시 pvs에 입력된 값",
                        "example": "/Mobius/Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ae",
                        "in": "path",
                        "description": "AE resourceName, AE 리소스의 이름",
                        "example": "Test_AE_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "acp",
                        "in": "path",
                        "description": "ACP resourceName, ACP 리소스의 이름",
                        "example": "Test_ACP_01",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "AE 하위 ACP 리소스 삭제",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ACP_Create": {
                "type": "object",
                "properties": {
                    "rn": {
                        "type": "string",
                        "description": "resourceName, 리소스의 이름",
                        "example": "Test_ACP_01"
                    },
                    "pv": {
                        "type": "object",
                        "example": {
                            "acr": [
                                {
                                    "acor": [
                                        "/Mobius/Test_AE_01"
                                    ],
                                    "acop": 63
                                }
                            ]
                        },
                        "description": "privileges, 리소스 접근 권한에 대한 정의"
                    },
                    "pvs": {
                        "type": "object",
                        "example": {
                            "acr": [
                                {
                                    "acor": [
                                        "/Mobius/Test_AE_01"
                                    ],
                                    "acop": 63
                                }
                            ]
                        },
                        "description": "selfPrivileges, ACP 리소스 자신의 접근 권한에 대한 정의"
                    }
                },
                "required": [
                    "rn",
                    "pv",
                    "pvs"
                ]
            },
            "ACP_Update": {
                "type": "object",
                "properties": {
                    "pv": {
                        "type": "object",
                        "example": {
                            "acr": [
                                {
                                    "acor": [
                                        "/Mobius/Test_AE_01"
                                    ],
                                    "acop": 63
                                }
                            ]
                        },
                        "description": "privileges, 리소스 접근 권한에 대한 정의"
                    },
                    "pvs": {
                        "type": "object",
                        "example": {
                            "acr": [
                                {
                                    "acor": [
                                        "/Mobius/Test_AE_01"
                                    ],
                                    "acop": 63
                                }
                            ]
                        },
                        "description": "selfPrivileges, ACP 리소스 자신의 접근 권한에 대한 정의"
                    }
                }
            },
...
        }
    }
}